requirements.txt
langchain==0.1.14
langchain-community==0.0.34
langchain-huggingface==0.0.42
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
sentence-transformers==2.4.1
pypdf==4.2.0
streamlit==1.37.0
fastapi==0.110.0
uvicorn==0.27.0
transformers==4.37.1
accelerate==0.28.0
mlflow==2.13.0
python-dotenv==1.0.1
torch==2.2.0
numpy==1.26.4
